+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.8.5
+
+* added `time_as_int` and `time_as_ms`